+++ doc/versionHistory.rst
@@ -2,6 +2,18 @@
 Version History
 ===============
 
+v.0.20.0
+--------
+
+* Add automatic refocus functionality in calculate_corrections. If not enough wavefront errors are estimated, it will attempt to use donut radii to focus the telescope.
+
+* Fix log message in `run_closed_loop` when rotator or elevation jump exceed the thresholds.
+
+v0.19.1
+-------
+
+* Add heartbeat and summaryState to remotes in `mtaos.py` and fix typo.
+
 v0.19.0
 -------